+++ odbcsv/Changelog.md
@@ -2,6 +2,8 @@
 
 ## 0.3.7 (next)
 
+* Fix: Allocated buffer sizes, now account for multi byte characters.
+
 ## 0.3.6
 
 * Fix: There has been an integer overflow causing a panic if an ODBC API call generated more than 2 ^ 15 warnings at once.